# Fast path for ascii user agents shorter than 1024 characters.