--- a/uasurfer.go
+++ b/uasurfer.go
@@ -162,7 +162,7 @@ func ParseUserAgent(ua string, dest *UserAgent) {
 }
 
 func parse(ua string, dest *UserAgent) {
-	ua = strings.ToLower(ua)
+	ua = normalise(ua)
 	switch {
 	case len(ua) == 0:
 		dest.OS.Platform = PlatformUnknown
@@ -177,4 +177,37 @@ func parse(ua string, dest *UserAgent) {
 		dest.evalBrowserVersion(ua)
 		dest.evalDevice(ua)
 	}
+}
+
+func normalise(ua string) string {
+	if len(ua) <= 1024 {
+		var buf [1024]byte
+		ascii := copyLower(buf[:len(ua)], ua)
+		if !ascii {
+			// Fall back for non ascii characters
+			return strings.ToLower(ua)
+		}
+		return string(buf[:len(ua)])
+	}
+	// Fallback for unusually long strings
+	return strings.ToLower(ua)
+}
+
+// copyLower copies a lowercase version of s to b. It assumes s contains only single byte characters
+// and will panic if b is nil or is not long enough to contain all the bytes from s.
+// It returns early with false if any chacracters were non ascii.
+func copyLower(b []byte, s string) bool {
+	for j := 0; j < len(s); j++ {
+		c := s[j]
+		if c > 127 {
+			return false
+		}
+
+		if 'A' <= c && c <= 'Z' {
+			c += 'a' - 'A'
+		}
+
+		b[j] = c
+	}
+	return true
 }