Falls back to old method if not.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkAgentSurfer-4            5568          4918          -11.67%
BenchmarkAgentSurferReuse-4       5453          5412          -0.75%
BenchmarkEvalSystem-4             2942          3162          +7.48%
BenchmarkEvalBrowserName-4        725           774           +6.76%
BenchmarkEvalBrowserVersion-4     89.5          96.8          +8.16%
BenchmarkEvalDevice-4             824           1066          +29.37%
BenchmarkParseChromeMac-4         2711          1609          -40.65%
BenchmarkParseChromeWin-4         2532          1383          -45.38%
BenchmarkParseChromeAndroid-4     6376          4950          -22.37%
BenchmarkParseSafariMac-4         3626          2349          -35.22%
BenchmarkParseSafariiPad-4        3841          2478          -35.49%

benchmark                         old allocs     new allocs     delta
BenchmarkAgentSurfer-4            2              2              +0.00%
BenchmarkAgentSurferReuse-4       1              1              +0.00%
BenchmarkEvalSystem-4             0              0              +0.00%
BenchmarkEvalBrowserName-4        0              0              +0.00%
BenchmarkEvalBrowserVersion-4     0              0              +0.00%
BenchmarkEvalDevice-4             0              0              +0.00%
BenchmarkParseChromeMac-4         3              2              -33.33%
BenchmarkParseChromeWin-4         3              2              -33.33%
BenchmarkParseChromeAndroid-4     3              2              -33.33%
BenchmarkParseSafariMac-4         3              2              -33.33%
BenchmarkParseSafariiPad-4        3              2              -33.33%

benchmark                         old bytes     new bytes     delta
BenchmarkAgentSurfer-4            320           207           -35.31%
BenchmarkAgentSurferReuse-4       240           127           -47.08%
BenchmarkEvalSystem-4             0             0             +0.00%
BenchmarkEvalBrowserName-4        0             0             +0.00%
BenchmarkEvalBrowserVersion-4     0             0             +0.00%
BenchmarkEvalDevice-4             0             0             +0.00%
BenchmarkParseChromeMac-4         336           208           -38.10%
BenchmarkParseChromeWin-4         304           192           -36.84%
BenchmarkParseChromeAndroid-4     368           224           -39.13%
BenchmarkParseSafariMac-4         336           208           -38.10%
BenchmarkParseSafariiPad-4        336           208           -38.10%
```